+++ src/tasks/enrich.py
@@ -11,34 +11,39 @@ from tasks.transform import (
 
 
 def add_etablissement_data(
-    lf_sirets: pl.LazyFrame, siret_column: str, type_siret: str
+    lf_sirets: pl.LazyFrame,
+    lf_etablissements: pl.LazyFrame,
+    siret_column: str,
+    type_siret: str,
 ) -> pl.LazyFrame:
-    # Récupération des données SIRET titulaires préparées
-    lf_etablissement = pl.scan_parquet(SIRENE_DATA_DIR / "etablissements.parquet")
-
     # Pas besoin de garder les SIRET qui ne matchent pas dans ce df intermédiaire, puisqu'on
     # merge in fine avec le reste des données (how = inner)
     lf_sirets = lf_sirets.join(
-        lf_etablissement, how="inner", left_on=siret_column, right_on="siret"
+        lf_etablissements, how="inner", left_on=siret_column, right_on="siret"
     )
     lf_sirets = lf_sirets.rename(
         {
-            # "departement": f"{type_siret}_departement",
             "latitude": f"{type_siret}_latitude",
             "longitude": f"{type_siret}_longitude",
+            "commune_code": f"{type_siret}_commune_code",
+            "commune_nom": f"{type_siret}_commune_nom",
+            "departement_code": f"{type_siret}_departement_code",
+            "departement_nom": f"{type_siret}_departement_nom",
+            "region_code": f"{type_siret}_region_code",
+            "region_nom": f"{type_siret}_region_nom",
         }
     )
     return lf_sirets
 
 
 def add_unite_legale_data(
-    lf_sirets: pl.LazyFrame, siret_column: str, type_siret: str
+    lf_sirets: pl.LazyFrame,
+    unites_legales_lf: pl.LazyFrame,
+    siret_column: str,
+    type_siret: str,
 ) -> pl.LazyFrame:
     # Extraction du SIREN à partir du SIRET (9 premiers caractères)
     lf_sirets = lf_sirets.with_columns(pl.col(siret_column).str.head(9).alias("siren"))
-
-    # Récupération des données des unités légales issues du flow de preprocess
-    unites_legales_lf = pl.scan_parquet(SIRENE_DATA_DIR / "unites_legales.parquet")
 
     # Pas besoin de garder les SIRET qui ne matchent pas dans ce df intermédiaire, puisqu'on
     # merge in fine avec le reste des données
@@ -52,6 +57,10 @@ def add_unite_legale_data(
 
 @task(log_prints=True)
 def enrich_from_sirene(lf: pl.LazyFrame):
+    # Récupération des données SIRET/SIREN préparées dans sirene-preprocess()
+    lf_etablissements = pl.scan_parquet(SIRENE_DATA_DIR / "etablissements.parquet")
+    lf_unites_legales = pl.scan_parquet(SIRENE_DATA_DIR / "unites_legales.parquet")
+
     # DONNÉES SIRENE ACHETEURS
 
     print("Extraction des SIRET des acheteurs...")
@@ -59,12 +68,15 @@ def enrich_from_sirene(lf: pl.LazyFrame):
 
     print("Ajout des données établissements (acheteurs)...")
     lf_sirets_acheteurs = add_etablissement_data(
-        lf_sirets_acheteurs, "acheteur_id", "acheteur"
+        lf_sirets_acheteurs, lf_etablissements, "acheteur_id", "acheteur"
     )
 
     print("Ajout des données unités légales (acheteurs)...")
     lf_sirets_acheteurs = add_unite_legale_data(
-        lf_sirets_acheteurs, siret_column="acheteur_id", type_siret="acheteur"
+        lf_sirets_acheteurs,
+        lf_unites_legales,
+        siret_column="acheteur_id",
+        type_siret="acheteur",
     )
 
     lf = lf.join(lf_sirets_acheteurs, how="left", on="acheteur_id")
@@ -81,12 +93,15 @@ def enrich_from_sirene(lf: pl.LazyFrame):
 
     print("Ajout des données établissements (titulaires)...")
     lf_sirets_titulaires = add_etablissement_data(
-        lf_sirets_titulaires, "titulaire_id", "titulaire"
+        lf_sirets_titulaires, lf_etablissements, "titulaire_id", "titulaire"
     )
 
     print("Ajout des données unités légales (titulaires)...")
     lf_sirets_titulaires = add_unite_legale_data(
-        lf_sirets_titulaires, siret_column="titulaire_id", type_siret="titulaire"
+        lf_sirets_titulaires,
+        lf_unites_legales,
+        siret_column="titulaire_id",
+        type_siret="titulaire",
     )
 
     # En joignant en utilisant à la fois le SIRET et le typeIdentifiant, on s'assure qu'on ne joint pas sur