+++ src/tasks/get.py
@@ -1,32 +1,40 @@
+import concurrent.futures
+import datetime
 import tempfile
 from collections.abc import Iterator
 from functools import partial
 from pathlib import Path
+from time import sleep
 
 import ijson
 import orjson
 import polars as pl
-from httpx import HTTPStatusError, TimeoutException, get, stream
+from httpx import Client, HTTPStatusError, TimeoutException, get
 from lxml import etree, html
-from lxml.html import HtmlElement
 from prefect import task
+from prefect.transactions import transaction
 
 from config import (
+    CACHE_EXPIRATION_TIME_HOURS,
     DECP_FORMAT_2022,
     DECP_FORMATS,
     DECP_PROCESSING_PUBLISH,
     DIST_DIR,
+    HTTP_CLIENT,
+    HTTP_HEADERS,
     DecpFormat,
 )
-from tasks.clean import clean_invalid_characters, extract_innermost_struct
+from tasks.clean import clean_decp, clean_invalid_characters, extract_innermost_struct
 from tasks.output import sink_to_files
-from tasks.utils import gen_artifact_row, stream_replace_bytestring
+from tasks.utils import gen_artifact_row, get_clean_cache_key, stream_replace_bytestring
 
 
 @task(retries=3, retry_delay_seconds=3)
 def stream_get(url: str, chunk_size=1024**2):  # chunk_size en octets (1 Mo par défaut)
     if url.startswith("http"):
-        with stream("GET", url, follow_redirects=True) as response:
+        with HTTP_CLIENT.stream(
+            "GET", url, headers=HTTP_HEADERS, follow_redirects=True
+        ) as response:
             yield from response.iter_bytes(chunk_size)
     else:
         # Données de test.
@@ -327,43 +335,97 @@ def norm_titulaire(titulaire: dict):
     return titulaire
 
 
-def get_etablissements(processed_parquet_path: Path):
+def get_etablissements() -> pl.LazyFrame:
     schema = {
         "siret": pl.String,
-        # "codeCommuneEtablissement": pl.String,
+        "codeCommuneEtablissement": pl.String,
         "latitude": pl.Float64,
         "longitude": pl.Float64,
-        # "etatAdministratifEtablissement": pl.Enum(["A", "F"]),
-        # "activitePrincipaleEtablissement": pl.String,
-        # "nomenclatureActivitePrincipaleEtablissement": pl.Enum(
-        #     ["NAF1993", "NAFRev1", "NAFRev2", "NAP"]
-        # ),
     }
 
-    lfs = []
     columns = list(schema.keys())
     print(columns)
+
     base_url = "https://files.data.gouv.fr/geo-sirene/last/dep/"
     htmlpage: str = get(base_url).text
-    htmlpage: HtmlElement = html.fromstring(htmlpage)
+    htmlpage: html.HtmlElement = html.fromstring(htmlpage)
+    http_client = Client()
+
+    # Préparation des hrefs
+    hrefs = []
     for link in htmlpage.findall(".//a"):
         href = link.get("href")
         if href.startswith("geo_siret"):
-            print(href)
-            href = base_url + href
-            try:
-                response = get(href, timeout=10).raise_for_status()
-            except (HTTPStatusError, TimeoutException) as e:
-                print(e)
-                print("Nouvel essai...")
-                response = get(href, timeout=10).raise_for_status()
-            content = response.content
-            lf = pl.scan_csv(content, infer_schema_length=50000)
-            lf = lf.select(columns)
-            lf = lf.with_columns(
-                [pl.col(col).cast(schema[col]).alias(col) for col in schema.keys()]
-            )
-            lfs.append(lf)
+            hrefs.append(base_url + href)
 
-    df_etablissements: pl.LazyFrame = pl.concat(lfs)
-    df_etablissements.sink_parquet(processed_parquet_path)
+    # Fonction de traitement pour un fichier
+    def process_file(_href: str):
+        print(_href.split("/")[-1])
+        try:
+            response = http_client.get(
+                _href, headers=HTTP_HEADERS, timeout=10
+            ).raise_for_status()
+        except (HTTPStatusError, TimeoutException) as err:
+            print(err)
+            print("Nouvel essai...")
+            response = http_client.get(
+                _href, headers=HTTP_HEADERS, timeout=10
+            ).raise_for_status()
+
+        content = response.content
+        lff = pl.scan_csv(content, infer_schema_length=1000, schema_overrides=schema)
+        lff = lff.select(columns)
+        lff = lff.with_columns(
+            [
+                pl.col("codeCommuneEtablissement").str.pad_start(5, "0"),
+                pl.col("siret").str.pad_start(14, "0"),
+            ]
+        )
+        return lff
+
+    # Traitement en parrallèle avec 8 threads
+    lfs = []
+    with concurrent.futures.ThreadPoolExecutor(max_workers=8) as executor:
+        futures = [executor.submit(process_file, href) for href in hrefs]
+        for future in concurrent.futures.as_completed(futures):
+            try:
+                lf = future.result()
+                lfs.append(lf)
+            except Exception as e:
+                print(f"Error processing file: {e}")
+
+    print("Concaténation...")
+    lf_etablissements: pl.LazyFrame = pl.concat(lfs)
+    return lf_etablissements
+
+
+def get_insee_cog_data(url, schema_overrides, columns) -> pl.DataFrame:
+    try:
+        df_insee = pl.read_csv(url, schema_overrides=schema_overrides, columns=columns)
+    except ConnectionResetError:
+        print("Connection error, retrying in 2 seconds...")
+        sleep(2)
+        df_insee = get_insee_cog_data(
+            url, schema_overrides=schema_overrides, columns=columns
+        )
+    return df_insee
+
+
+@task(
+    log_prints=True,
+    persist_result=True,
+    cache_expiration=datetime.timedelta(hours=CACHE_EXPIRATION_TIME_HOURS),
+    cache_key_fn=get_clean_cache_key,
+)
+def get_clean(resource, resources_artifact: list) -> pl.DataFrame or None:
+    # Récupération des données source...
+    with transaction():
+        lf, decp_format = get_resource(resource, resources_artifact)
+
+        # Nettoyage des données source et typage des colonnes...
+        # si la ressource est dans un format supporté
+        if lf is not None:
+            lf = clean_decp(lf, decp_format)
+            df = lf.collect(engine="streaming")
+
+    return df